Joind.in Automation
This is the first and very simple step.
Somewhere in the beginning of the month this cronjob should be executed, it will find this month's meeting and will create a simple placeholder at joind.in.
Since joind.in requires approval it will just link you to a search query that will be valid soon.

Next steps may be:
- run another cron every day until the event is found and try to create talk data
- add a url that can figure out the link to this month's talk to list them on the slide